+++ src/helpers/utilities.ts
@@ -89,9 +89,9 @@ export const updatePrecisionToDisplay = (
   amount: BigNumberish,
   nativePrice: BigNumberish,
   roundUp: boolean = false
-): BigNumberish => {
-  if (!amount) return 0;
-  if (!nativePrice) return amount;
+): string => {
+  if (!amount) return '0';
+  if (!nativePrice) return new BigNumber(amount).toFixed();
   const roundingMode = roundUp ? BigNumber.ROUND_UP : BigNumber.ROUND_DOWN;
   const bnAmount = new BigNumber(amount);
   const significantDigitsOfNativePriceInteger = new BigNumber(nativePrice)
@@ -207,11 +207,7 @@ export const handleSignificantDecimals = (
   buffer: number = 3
 ): string => {
   if (lessThan(new BigNumber(value).abs(), 1)) {
-    decimals =
-      new BigNumber(value)
-        .toFixed()
-        .slice(2)
-        .search(/[^0]/g) + buffer;
+    decimals = new BigNumber(value).toFixed().slice(2).search(/[^0]/g) + buffer;
     decimals = Math.min(decimals, 8);
   } else {
     decimals = Math.min(decimals, buffer);