+++ src/redux/reducers.js
@@ -1,6 +1,5 @@
 import { combineReducers } from 'redux';
 
-import actionSheetManager from './actionSheetManager';
 import addCash from './addCash';
 import appState from './appState';
 import charts from './charts';
@@ -13,7 +12,6 @@ import gas from './gas';
 import imageMetadata from './imageMetadata';
 import keyboardHeight from './keyboardHeight';
 import multicall from './multicall';
-import nonce from './nonce';
 import openStateSettings, {
   openSavingsReducer,
   openSmallBalancesReducer,
@@ -22,6 +20,7 @@ import raps from './raps';
 import requests from './requests';
 import settings from './settings';
 import showcaseTokens from './showcaseTokens';
+import swap from './swap';
 import uniqueTokens from './uniqueTokens';
 import uniswap from './uniswap';
 import uniswapLiquidity from './uniswapLiquidity';
@@ -29,7 +28,6 @@ import walletconnect from './walletconnect';
 import wallets from './wallets';
 
 export default combineReducers({
-  actionSheetManager,
   addCash,
   appState,
   charts,
@@ -42,7 +40,6 @@ export default combineReducers({
   imageMetadata,
   keyboardHeight,
   multicall,
-  nonce,
   openSavings: openSavingsReducer,
   openSmallBalances: openSmallBalancesReducer,
   openStateSettings,
@@ -50,6 +47,7 @@ export default combineReducers({
   requests,
   settings,
   showcaseTokens,
+  swap,
   uniqueTokens,
   uniswap,
   uniswapLiquidity,