+++ src/screens/CurrencySelectModal.js
@@ -1,6 +1,6 @@
 import { useIsFocused, useRoute } from '@react-navigation/native';
 import analytics from '@segment/analytics-react-native';
-import { concat, map, toLower } from 'lodash';
+import { map, toLower } from 'lodash';
 import matchSorter from 'match-sorter';
 import React, {
   Fragment,
@@ -23,11 +23,9 @@ import {
 } from '../components/exchange';
 import { Column, KeyboardFixedOpenLayout } from '../components/layout';
 import { Modal } from '../components/modal';
-import { addHexPrefix } from '../handlers/web3';
-import CurrencySelectionTypes from '../helpers/currencySelectionTypes';
-import tokenSectionTypes from '../helpers/tokenSectionTypes';
-import { delayNext } from '../hooks/useMagicAutofocus';
-import { useNavigation } from '../navigation/Navigation';
+import { addHexPrefix } from '@rainbow-me/handlers/web3';
+import CurrencySelectionTypes from '@rainbow-me/helpers/currencySelectionTypes';
+import tokenSectionTypes from '@rainbow-me/helpers/tokenSectionTypes';
 import {
   useInteraction,
   useMagicAutofocus,
@@ -36,9 +34,11 @@ import {
   useUniswapAssets,
   useUniswapAssetsInWallet,
 } from '@rainbow-me/hooks';
+import { delayNext } from '@rainbow-me/hooks/useMagicAutofocus';
+import { useNavigation } from '@rainbow-me/navigation/Navigation';
 import Routes from '@rainbow-me/routes';
-import { position } from '@rainbow-me/styles';
-import { filterList, filterScams } from '@rainbow-me/utils';
+import { colors, position } from '@rainbow-me/styles';
+import { filterList } from '@rainbow-me/utils';
 
 const TabTransitionAnimation = styled(Animated.View)`
   ${position.size('100%')};
@@ -68,7 +68,6 @@ export default function CurrencySelectModal() {
   const { navigate, dangerouslyGetState } = useNavigation();
   const {
     params: {
-      category,
       onSelectCurrency,
       restoreFocusOnSwapModal,
       setPointerEvents,
@@ -90,10 +89,11 @@ export default function CurrencySelectModal() {
   ]);
 
   const {
-    curatedAssets,
+    curatedNotFavorited,
     favorites,
     globalHighLiquidityAssets,
     globalLowLiquidityAssets,
+    globalVerifiedHighLiquidityAssets,
     loadingAllTokens,
     updateFavorites,
   } = useUniswapAssets();
@@ -111,43 +111,57 @@ export default function CurrencySelectModal() {
         filteredList = headerlessSection(filteredList);
       }
     } else if (type === CurrencySelectionTypes.output) {
-      const curatedSection = concat(favorites, curatedAssets);
       if (searchQueryForSearch) {
-        const [filteredBest, filteredHigh, filteredLow] = map(
-          [curatedSection, globalHighLiquidityAssets, globalLowLiquidityAssets],
+        const [
+          filteredFavorite,
+          filteredVerified,
+          filteredHighUnverified,
+          filteredLow,
+        ] = map(
+          [
+            favorites,
+            globalVerifiedHighLiquidityAssets,
+            globalHighLiquidityAssets,
+            globalLowLiquidityAssets,
+          ],
           section => searchCurrencyList(section, searchQueryForSearch)
         );
 
         filteredList = [];
-        filteredBest.length &&
+        filteredFavorite.length &&
           filteredList.push({
-            data: filteredBest,
+            color: colors.yellowFavorite,
+            data: filteredFavorite,
+            title: tokenSectionTypes.favoriteTokenSection,
+          });
+
+        filteredVerified.length &&
+          filteredList.push({
+            data: filteredVerified,
             title: tokenSectionTypes.verifiedTokenSection,
             useGradientText: IS_TESTING === 'true' ? false : true,
           });
 
-        const filteredHighWithoutScams = filterScams(
-          filteredBest,
-          filteredHigh
-        );
-
-        filteredHighWithoutScams.length &&
+        filteredHighUnverified.length &&
           filteredList.push({
-            data: filteredHighWithoutScams,
+            data: filteredHighUnverified,
             title: tokenSectionTypes.unverifiedTokenSection,
           });
 
-        const filteredLowWithoutScams = filterScams(filteredBest, filteredLow);
-
-        filteredLowWithoutScams.length &&
+        filteredLow.length &&
           filteredList.push({
-            data: filteredLowWithoutScams,
+            data: filteredLow,
             title: tokenSectionTypes.lowLiquidityTokenSection,
           });
       } else {
         filteredList = [
           {
-            data: concat(favorites, curatedAssets),
+            color: colors.yellowFavorite,
+            data: favorites,
+            title: tokenSectionTypes.favoriteTokenSection,
+          },
+          {
+            data: curatedNotFavorited,
             title: tokenSectionTypes.verifiedTokenSection,
             useGradientText: IS_TESTING === 'true' ? false : true,
           },
@@ -157,8 +171,9 @@ export default function CurrencySelectModal() {
     setIsSearching(false);
     return filteredList;
   }, [
-    curatedAssets,
+    curatedNotFavorited,
     favorites,
+    globalVerifiedHighLiquidityAssets,
     globalHighLiquidityAssets,
     globalLowLiquidityAssets,
     searchQueryForSearch,
@@ -185,7 +200,6 @@ export default function CurrencySelectModal() {
         [asset.address]: isFavorited,
       }));
       analytics.track('Toggled an asset as Favorited', {
-        category,
         isFavorited,
         name: asset.name,
         symbol: asset.symbol,
@@ -193,7 +207,7 @@ export default function CurrencySelectModal() {
         type,
       });
     },
-    [category, type]
+    [type]
   );
 
   const handleSelectAsset = useCallback(
@@ -202,7 +216,6 @@ export default function CurrencySelectModal() {
       onSelectCurrency(item);
       if (searchQueryForSearch) {
         analytics.track('Selected a search result in Swap', {
-          category,
           name: item.name,
           searchQueryForSearch,
           symbol: item.symbol,
@@ -220,7 +233,6 @@ export default function CurrencySelectModal() {
       searchQueryForSearch,
       dangerouslyGetState,
       navigate,
-      category,
       type,
     ]
   );