+++ CHANGELOG.md
@@ -12,6 +12,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Removed
 
+## [1.2.48](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.48)
+### Changed
+* UI Bug Fixes
+
+## [1.2.47](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.47)
+### Changed
+* Search Improvements
+* WalletConnect Error Reporting
+
+## [1.2.46](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.46)
+### Changed
+* Token Approval Improvements
+
+## [1.2.45](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.45)
+### Changed
+* WC nonce fix
+
+## [1.2.44](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.44)
+### Changed
+* Improve nonce tracking
+* Fix Wallet Connect modals for 0 ETH wallets
+
 ## [1.2.43](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.43)
 ### Added
 * Speed up & cancel transactions